Et hop, chasse et tannerie opérationnels !

+++ Systems/PlayerSystem/PlayerSystem.cs
@@ -965,19 +965,34 @@ namespace NWN.Systems
 
       return 0;
     }
+    private static int HandleBeforeAcquireItem(uint oidSelf)
+    {
+      if (NWScript.GetTag(NWScript.StringToObject(EventsPlugin.GetEventData("ITEM"))) == "undroppable_item")
+        EventsPlugin.SkipEvent();
+      
+      return 0;
+    }
     private static int HandlePCAcquireItem(uint oidSelf)
     {
       var oPC = NWScript.GetModuleItemAcquiredBy();
       //Console.WriteLine(NWScript.GetName(oPC));
-      //Console.WriteLine(NWScript.GetName(NWScript.GetArea(oPC)));
 
       if (Convert.ToBoolean(NWScript.GetIsPC(oPC)))
       {
         var oItem = NWScript.GetModuleItemAcquired();
         var oAcquiredFrom = NWScript.GetModuleItemAcquiredFrom();
 
+        //Console.WriteLine(NWScript.GetTag(oItem));
+
         if (Convert.ToBoolean(NWScript.GetIsObjectValid(oItem)))
         {
+          if(NWScript.GetTag(oItem) == "undroppable_item")
+          {
+            NWScript.CopyObject(oItem, NWScript.GetLocation(oAcquiredFrom), oAcquiredFrom);
+            NWScript.DestroyObject(oItem);
+            return 0;
+          }
+
           if (NWScript.GetTag(oItem) == "item_pccorpse" && NWScript.GetTag(oAcquiredFrom) == "pccorpse_bodybag")
           {
             DeletePlayerCorpseFromDatabase(NWScript.GetLocalInt(oItem, "_PC_ID"));
@@ -995,7 +1010,6 @@ namespace NWN.Systems
               }
               oCorpse = NWScript.GetObjectByTag("pccorpse", i++);
             }
-
           }
           /*En pause jusqu'à ce que le système de transport soit en place
           if (NWScript.GetMovementRate(oPC) != CreaturePlugin.NWNX_CREATURE_MOVEMENT_RATE_IMMOBILE)
@@ -1057,9 +1071,7 @@ namespace NWN.Systems
           }
         }
 
-        //EN FONCTION DE SI LA ZONE EST REST OU PAS, ON AFFICHE LA PROGRESSION DU JOURNAL DE CRAFT
-        Player player;
-        if (Players.TryGetValue(oidSelf, out player))
+        if (Players.TryGetValue(oidSelf, out Player player)) //EN FONCTION DE SI LA ZONE EST REST OU PAS, ON AFFICHE LA PROGRESSION DU JOURNAL DE CRAFT
         {
           player.previousArea = oArea;
 
@@ -1101,6 +1113,18 @@ namespace NWN.Systems
 
             if (nbPlayersInArea == 0)
               NWScript.AssignCommand(area.oid, () => NWScript.DelayCommand(1500.0f, () => area.Clean())); // 25 minutes
+
+            if(area.tag == $"entrepotpersonnel_{NWScript.GetName(player.oid)}")
+            {
+              uint storageToSave = NWScript.GetFirstObjectInArea(NWScript.GetArea(player.oid));
+              if (NWScript.GetTag(storageToSave) != "ps_entrepot")
+                storageToSave = NWScript.GetNearestObjectByTag("ps_entrepot", storageToSave);
+
+              var saveStorage = NWScript.SqlPrepareQueryCampaign(ModuleSystem.database, $"UPDATE playerCharacters set storage = @storage where rowid = @characterId");
+              NWScript.SqlBindInt(saveStorage, "@characterId", player.characterId);
+              NWScript.SqlBindObject(saveStorage, "@storage", storageToSave);
+              NWScript.SqlStep(saveStorage);
+            }
           }
         }
       }
